Refactor temp code
